requirements.txt
streamlit==1.28.1
pandas==2.1.4
plotly==5.17.0
sqlalchemy==2.0.23
pymysql==1.1.0
python-dotenv==1.0.0
numpy==1.26.2
